client/requirements.txt
python-socketio==5.13.0
websocket-client
pillow
google-genai
pydantic
python-dotenv
py2app
macos-notifications